Testing workaround for compilation problems on weird mingw installations.

+++ src/pilrc/pilrc.h
@@ -41,10 +41,11 @@
 #ifndef _pilrc_h
 #define _pilrc_h                                 // RMA : multiples include protection
 
+#include <stdint.h>
 //#if (SIZEOF_INT == SIZEOF_CHAR_P)
 //typedef int p_int;
 //#elif (SIZEOF_LONG == SIZEOF_CHAR_P)
-typedef long p_int;
+typedef intptr_t p_int;
 //#endif
 
 #ifndef strdup
